+++ src/modelldcatnotordf/__init__.py
@@ -12,3 +12,5 @@ try:
     __version__ = version(__name__)
 except PackageNotFoundError:  # pragma: no cover
     __version__ = "unknown"
+
+from modelldcatnotordf.modelldcatno import ModelProperty
